backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
python-decouple==3.4
email-validator==1.1.3
boto3==1.18.44
tensorflow==2.6.0
numpy==1.19.5
pandas==1.3.3
requests==2.26.0
aiohttp==3.7.4
python-socketio==5.4.0
web3==5.23.1
redis==3.5.3
prometheus-client==0.11.0
sentry-sdk==1.3.1